+++ src/Decoder.ts
@@ -264,15 +264,15 @@ export const partial = <A>(
  * @category combinators
  * @since 2.2.7
  */
-export const karray = <I, A>(item: Decoder<I, A>): Decoder<Array<I>, Array<A>> =>
-  K.array(M)((i, e) => FS.of(DE.index(i, DE.optional, e)))(item)
+export const items = <I, A>(item: Decoder<I, A>): Decoder<Array<I>, Array<A>> =>
+  K.items(M)((i, e) => FS.of(DE.index(i, DE.optional, e)))(item)
 
 /**
  * @category combinators
  * @since 2.2.8
  */
 export const array = <A>(item: Decoder<unknown, A>): Decoder<unknown, Array<A>> =>
-  pipe(UnknownArray, compose(karray(item)))
+  pipe(UnknownArray, compose(items(item)))
 
 /**
  * @category combinators