+++ CHANGELOG.md
@@ -30,12 +30,13 @@
     - add `id` (@gcanti)
     - drop `type` in favour of `props` (@gcanti)
     - drop `partial` in favour of `partialProps` (@gcanti)
+    - rename `array` to `items` (@gcanti)
   - `KleisliDecoder`
     - remove old `Decoder.ts` module and rename `KleisliDecoder` to `Decoder` (@gcanti)
     - add `string`, `number`, `boolean`, `UnknownArray`, `UnknownRecord` primitives (@gcanti)
     - drop `type` in favour of `props` (@gcanti)
     - drop `partial` in favour of `partialProps` (@gcanti)
-    - rename `array` to `karray`
+    - rename `array` to `items`
     - rename `record` to `krecord`
     - rename `tuple` to `ktuple`
     - rename `sum` to `ksum`
@@ -49,7 +50,7 @@
     - add `string`, `number`, `boolean`, `UnknownArray`, `UnknownRecord` primitives (@gcanti)
     - drop `type` in favour of `props` (@gcanti)
     - drop `partial` in favour of `partialProps` (@gcanti)
-    - rename `array` to `karray`
+    - rename `array` to `items`
     - rename `record` to `krecord`
     - rename `tuple` to `ktuple`
     - rename `sum` to `ksum`